Release 0.25.1
Fixes the circular dev-dependency between edgefirst-codec and
edgefirst-image that required a sed workaround in the release workflow.
The bench_nvjpeg benchmark is moved to edgefirst-image where it belongs
(it needs ImageProcessor to allocate the CUDA-backed PBO destination,
and image already depends on codec). The release workflow sed hack and
the dev-dep entry are both removed.
