+++ Cargo.toml
@@ -15,7 +15,7 @@ authors = ["Au-Zone Technologies <support@au-zone.com>"]
 homepage = "https://edgefirst.ai"
 repository = "https://github.com/EdgeFirstAI/hal"
 license = "Apache-2.0"
-version = "0.25.0"
+version = "0.25.1"
 edition = "2021"
 readme = "README.md"
 
@@ -30,12 +30,12 @@ ctor = "0.4.2"
 dma-heap = "0.4.1"
 # Internal crates
 edgefirst-bench = { path = "crates/bench" }
-edgefirst-codec = { version = "0.25.0", path = "crates/codec" }
-edgefirst-hal = { version = "0.25.0", path = "crates/hal" }
-edgefirst-decoder = { version = "0.25.0", path = "crates/decoder" }
-edgefirst-image = { version = "0.25.0", path = "crates/image" }
-edgefirst-tensor = { version = "0.25.0", path = "crates/tensor" }
-edgefirst-tracker = { version = "0.25.0", path = "crates/tracker" }
+edgefirst-codec = { version = "0.25.1", path = "crates/codec" }
+edgefirst-hal = { version = "0.25.1", path = "crates/hal" }
+edgefirst-decoder = { version = "0.25.1", path = "crates/decoder" }
+edgefirst-image = { version = "0.25.1", path = "crates/image" }
+edgefirst-tensor = { version = "0.25.1", path = "crates/tensor" }
+edgefirst-tracker = { version = "0.25.1", path = "crates/tracker" }
 enum_dispatch = "0.3.13"
 env_logger = "0.11.10"
 fast-math = "0.1.1"
