+++ crates/codec/benches/codec_benchmark.rs
@@ -17,7 +17,7 @@
 
 use edgefirst_bench::{run_bench, BenchSuite};
 use edgefirst_codec::{ImageDecoder, ImageLoad};
-use edgefirst_tensor::{DType, PixelFormat, Tensor, TensorMemory};
+use edgefirst_tensor::{PixelFormat, Tensor, TensorMemory};
 use std::sync::LazyLock;
 
 const WARMUP: usize = 10;
@@ -310,70 +310,6 @@ fn bench_v4l2_persistent_stream(suite: &mut BenchSuite) {
 }
 
 // =============================================================================
-// 6. nvJPEG GPU decode (on-target only: Linux + CUDA + libnvjpeg)
-//
-// Decodes straight into a CUDA-registered PBO (what ImageProcessor::create_image
-// yields on Jetson), emitting GPU-resident RGB. Skips cleanly on any host
-// without nvJPEG/CUDA — bench_compare.py treats absent cells as one-sided and
-// never gates on them, so this is safe to keep in the default suite.
-//
-//   codec/jpeg/nvjpeg/rgbi/<fixture>  — decode-only into the PBO device pointer
-// =============================================================================
-
-fn bench_nvjpeg(suite: &mut BenchSuite) {
-    if !edgefirst_codec::nvjpeg_available() {
-        // nvJPEG is opt-in (off by default); also requires CUDA + libnvjpeg.
-        println!("\n== nvjpeg: SKIP (set EDGEFIRST_ENABLE_NVJPEG=1; needs CUDA + libnvjpeg) ==\n");
-        return;
-    }
-    println!("\n== edgefirst-codec: nvjpeg decode into PBO (GPU-resident RGB) ==\n");
-
-    let processor = match edgefirst_image::ImageProcessor::new() {
-        Ok(p) => p,
-        Err(e) => {
-            println!("nvjpeg bench SKIP: ImageProcessor unavailable: {e}");
-            return;
-        }
-    };
-
-    // (width, height, jpeg bytes, cell name)
-    let cases: [(usize, usize, &LazyLock<Vec<u8>>, &str); 2] = [
-        (1280, 720, &ZIDANE_JPG, "codec/jpeg/nvjpeg/rgbi/zidane_720p"),
-        (640, 640, &GIRAFFE_JPG, "codec/jpeg/nvjpeg/rgbi/giraffe_640"),
-    ];
-
-    for (w, h, jpeg, name) in cases {
-        // RGB PBO destination — auto-selected backend (PBO + CUDA on Jetson).
-        let mut tensor = match processor.create_image(w, h, PixelFormat::Rgb, DType::U8, None) {
-            Ok(t) => t,
-            Err(e) => {
-                println!("{name} SKIP: create_image failed: {e}");
-                continue;
-            }
-        };
-        let mut decoder = ImageDecoder::new();
-        // Warm once and confirm nvJPEG actually fired (RGB) rather than the CPU
-        // path silently producing NV12 into a non-CUDA tensor.
-        let info = tensor.load_image(&mut decoder, jpeg).unwrap();
-        if info.format != PixelFormat::Rgb {
-            println!(
-                "{name} SKIP: decoded to {:?}, not Rgb — nvJPEG did not engage \
-                 (destination not CUDA-backed?)",
-                info.format
-            );
-            continue;
-        }
-
-        let r = run_bench(name, WARMUP, ITERATIONS, || {
-            tensor.load_image(&mut decoder, jpeg).unwrap();
-            std::hint::black_box(&tensor);
-        });
-        r.print_summary();
-        suite.record(&r);
-    }
-}
-
-// =============================================================================
 // Main
 // =============================================================================
 
@@ -385,7 +321,6 @@ fn main() {
     bench_image_crate(&mut suite);
     bench_exif_overhead(&mut suite);
     bench_v4l2_persistent_stream(&mut suite);
-    bench_nvjpeg(&mut suite);
 
     suite.finish();
 }
