+++ crates/image/Cargo.toml
@@ -95,6 +95,10 @@ name = "decode_pipeline_benchmark"
 harness = false
 
 [[bench]]
+name = "nvjpeg_benchmark"
+harness = false
+
+[[bench]]
 name = "convert_matrix_benchmark"
 harness = false
